editor/spellchecker/tests/test_suggest.html

<!DOCTYPE html>
<html>
<head>
  <title>Test for nsIEditorSpellChecker.suggest</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>

<div contenteditable id="en-US" lang="en-US">missspelled</div>

<pre id="test">
<script class="testbody" type="text/javascript">
add_task(async function() {
  await new Promise(resolve => SimpleTest.waitForFocus(resolve));

  let { maybeOnSpellCheck } = SpecialPowers.ChromeUtils.importESModule(
    "resource://testing-common/AsyncSpellCheckTestHelper.sys.mjs"
  );

  let element = document.getElementById("en-US");
  element.focus();

  await new Promise(resolve => maybeOnSpellCheck(element, resolve));

  let editingSession = SpecialPowers.wrap(window).docShell.editingSession;
  let editor = editingSession.getEditorForWindow(window);
  let spellchecker = SpecialPowers.Cu.createSpellChecker();
  spellchecker.setFilterType(spellchecker.FILTERTYPE_NORMAL);
  await new Promise(resolve => spellchecker.InitSpellChecker(editor, false, resolve));

  let suggestions = await spellchecker.suggest("misspelled", 5);
  is(suggestions.length, 0, "\"misspelled\" is correct word");

  // Hunspell 1.7.3 switched suggestion timeouts from CPU time (clock()) to wall
  // time (std::chrono::steady_clock). On slower builds (sanitizers, 32-bit) the
  // 100ms limit is hit before all 5 suggestions are generated.
  suggestions = await spellchecker.suggest("missspelled", 5);
  ok(suggestions.length >= 3 && suggestions.length <= 5, "\"missspelled\" isn't correct word");
});
</script>
</pre>
</body>
</html>
